requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Computer Vision and MediaPipe
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.13

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# LLM Integration
openai>=1.3.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.10.0
flake8>=6.1.0

